% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllLabelledNewicksFromNewickNumberStrings.R
\name{AllLabelledNewicksFromNewickNumberStrings}
\alias{AllLabelledNewicksFromNewickNumberStrings}
\title{Generate All Labelled Newicks from Newick Number Strings}
\usage{
AllLabelledNewicksFromNewickNumberStrings(
  NewickNumberStrings,
  TipLabels = NULL
)
}
\arguments{
\item{NewickNumberStrings}{A vector of Newick number strings.}

\item{TipLabels}{A character vector of tip labels to use.}
}
\value{
A list, one for each input Newick number string, each containing all possible labelled Newicks.
}
\description{
Generates all possible labelled newick strings from a set of Newick number strings.
}
\examples{

# Get information for trees with N tips:
AllLabelledNewicksFromNewickNumberStrings(c("(1,(2,(3)));", "(2,(1,(3)));"))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
