% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllNewickNumbersByAddition.R
\name{AllNewickNumbersByAddition}
\alias{AllNewickNumbersByAddition}
\title{Generate All Unlabelled Trees Through Addition}
\usage{
AllNewickNumbersByAddition(N)
}
\arguments{
\item{N}{The required number of tips.}
}
\value{
A list of Newick number strings partitioned by number of internal nodes (1 to N - 1).
}
\description{
Generates all unlabelled rooted multifurcating trees of N tips.
}
\details{
Function works using notion in Felsenstein (2004) or simply starting with the smallest possible tree and additively generating trees up to N tips by adding a new tip in every possible place.

Note that currently this function fails to generate all possible trees past N = 7 so it is not recommended for use.
}
\examples{

# Generate all trees of five tips:
AllNewickNumbersByAddition(N = 5)

}
\references{
Felsenstein 2004
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{combinatorics,}
\keyword{phylogeny,}
\keyword{trees}
