% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayPosition2ArrayCoordinates.R
\name{ArrayPosition2ArrayCoordinates}
\alias{ArrayPosition2ArrayCoordinates}
\title{Get Array Coordinates From Array Position}
\usage{
ArrayPosition2ArrayCoordinates(Position, ArrayDimensions)
}
\arguments{
\item{Position}{The number of the Nth value in the array (an integer).}

\item{ArrayDimensions}{A vector of integers giving the dimensions of the array.}
}
\value{
A vector the same length as \code{ArrayDimensions} giving the coordinates.
}
\description{
Simply returns the coordinates for the Nth value in an array.
}
\details{
Depending on different criteria N tips will produce a specific number of trees.
}
\examples{

# Get the coordinates of the 17th value in a
# 2 by 4 by 6 array:
ArrayPosition2ArrayCoordinates(17, c(2,4,6))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{array,}
\keyword{coordinates}
