% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularVoronoi.R
\name{CircularVoronoi}
\alias{CircularVoronoi}
\title{Circular Voronoi Plot}
\usage{
CircularVoronoi(x, y, heat, EdgeFactor = 1.1)
}
\arguments{
\item{x}{The x coordinates of the points to plot.}

\item{y}{The y coordinates of the points to plot.}

\item{heat}{Numerical values that will determine the colour of each tile plotted (i.e., as a heat map).}

\item{EdgeFactor}{The proportional extension of the edge of the plotting circle (i.e., the default of 1.1 adds 10% to the dimatere of the point cloud).}
}
\value{
A circular Voronoi plot with tiles coloured according to heat values.
}
\description{
Makes a circular Voronoi plot.
}
\details{
Intended for visualising two-dimensional projections of hyperdimensional point clouds that approximate a hypersphere. Is effectively a circular extension of the Voronoi function in the \link{deldir} package.
}
\examples{

# Create some random data:
N <- 100
Radius <- runif(n = N, min = 0, max = 2)
Angle <- sample(1:360, size = N, replace = TRUE)
x <- Radius * cos(Angle * (pi / 180))
y <- Radius * sin(Angle * (pi / 180))
heat <- sort(runif(n = N))

# Create treespace tile plot:
CircularVoronoi(x, y, heat, EdgeFactor = 1.05)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{ancestor}
