% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllPartitionCombinations.R
\name{GetAllPartitionCombinations}
\alias{GetAllPartitionCombinations}
\title{All Partition Combinations}
\usage{
GetAllPartitionCombinations(PartitionSizes, Labels)
}
\arguments{
\item{PartitionSizes}{A numeric vector of partition sizes.}

\item{Labels}{A character vector of labels.}
}
\value{
A matrix of all possible label combinations.
}
\description{
Given a set of partition sizes and labels produces all possible combinations.
}
\examples{

# Get information for trees with N tips:
GetAllPartitionCombinations(c(1,2,3), LETTERS[1:6])

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
