% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNLabelledMultifurcatingRootedTrees.R
\name{GetNLabelledMultifurcatingRootedTrees}
\alias{GetNLabelledMultifurcatingRootedTrees}
\title{Get N Labelled Multifucrating Rooted Trees}
\usage{
GetNLabelledMultifurcatingRootedTrees(N)
}
\arguments{
\item{N}{The number of tips.}
}
\value{
A list containg two items, TotalTrees (the total number of trees) and TreesByNodeCount (a vector of trees by node count from 1 to N - 1).
}
\description{
Gets the total number of labelled multifurcating rooted trees for N tips.
}
\details{
Uses equation 3.1 (and illustrated in Figure 3.6) of Felsenstein (2004).

Note that direct counts are possible only up to N = 145, beyond that R will return only "Inf".
}
\examples{

# Get numbers of trees:
GetNLabelledMultifurcatingRootedTrees(8)

# Get just the total number of trees for 1:10 tips
unlist(lapply(as.list(1:10), function(x)
  GetNLabelledMultifurcatingRootedTrees(x)$TotalTrees))

}
\references{
Felsenstein 2004
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{combinatorics,}
\keyword{phylogeny,}
\keyword{trees}
