% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNTrees.R
\name{GetNTrees}
\alias{GetNTrees}
\title{Get N Trees}
\usage{
GetNTrees(N)
}
\arguments{
\item{N}{The number of tips.}
}
\value{
{The number of trees.}
}
\description{
Gets the total number of trees for N tips.
}
\details{
Depeding on different criteria N tips will produce a specific number of trees.
}
\examples{

# Create some example coordinates:
N <- 3

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{trees}
