% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypRspace-package.R
\docType{package}
\name{hypRspace-package}
\alias{hypRspace-package}
\alias{hypRspace}
\title{Biological Hyperspace Functions}
\description{
Functions for biological hyperspaces.
}
\examples{

# Nothing yet.

}
\references{
Nothing yet.
}
\author{
Graeme T. Lloyd <graemetlloyd@gmail.com>
}
\keyword{biology,hyperspace}
