% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LadderizeNewick.R
\name{LadderizeNewickNumberString}
\alias{LadderizeNewickNumberString}
\title{Ladderize Newick Number String}
\usage{
LadderizeNewickNumberString(NewickNumberString)
}
\arguments{
\item{NewickNumberString}{A split Newick number string.}
}
\value{
A vector of labelled Newick strings corresponding to unique trees.
}
\description{
Takes as input Newick number string and returns the left ladderized version.
}
\examples{

# Get information for trees with N tips:
LadderizeNewickNumberString(NewickNumberStringSplitter("((4),(1,(2)));"))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
