% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFakeLabels.R
\name{MakeFakeLabels}
\alias{MakeFakeLabels}
\title{Make fake labels}
\usage{
MakeFakeLabels(N)
}
\arguments{
\item{N}{The number of labels required.}
}
\value{
A vector of labels (character strings).
}
\description{
Generates a set of fake labels (character strings).
}
\examples{

# Get information for trees with N tips:
MakeFakeLabels(N = 100)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
