% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewickNumberStringSplitter.R
\name{NewickNumberStringSplitter}
\alias{NewickNumberStringSplitter}
\title{Split Newick Number String}
\usage{
NewickNumberStringSplitter(NewickNumberString)
}
\arguments{
\item{NewickNumberString}{A vector of Newick Number Strings.}
}
\value{
A vector containing each component of a Newick number string.
}
\description{
Takes a Newick number string and splits it into its' component parts.
}
\details{
Prentheeses, commas, labels and semi-colon.
}
\examples{

# Get information for trees with N tips:
NewickNumberStringSplitter("(1,(2,(3)));")

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
