% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewickStringSplitter.R
\name{NewickStringSplitter}
\alias{NewickStringSplitter}
\title{Split Newick String}
\usage{
NewickStringSplitter(NewickString)
}
\arguments{
\item{NewickString}{A vector of Newick strings.}
}
\value{
A vector containing each component of a Newick string.
}
\description{
Takes a Newick string and splits it into its' component parts.
}
\details{
Parentheeses, commas, labels and semi-colon.
}
\examples{

# Get information for trees with N tips:
NewickStringSplitter("(Aa,(Bb,Cc,(Dd,Ee,Ff)));")

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
