% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueNewickNumberStrings.R
\name{UniqueNewickNumberStrings}
\alias{UniqueNewickNumberStrings}
\title{Get Unique Newick Number Strings}
\usage{
UniqueNewickNumberStrings(NewickNumberStrings)
}
\arguments{
\item{NewickNumberStrings}{A vector of labelled Newick strings.}
}
\value{
A vector of Newick number strings corresponding to unique treeshapes.
}
\description{
Takes as input a vector of Newick number strings and returns only those pertaining to unique treeshapes.
}
\details{
Principle of free rotation.
}
\examples{

# Get information for trees with N tips:
UniqueNewickNumberStrings(c("(1,(2),(2));", "(1,(2),(2));"))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
