% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueNewickStrings.R
\name{UniqueNewickStrings}
\alias{UniqueNewickStrings}
\title{Get Unique Newick Strings}
\usage{
UniqueNewickStrings(NewickStrings)
}
\arguments{
\item{NewickStrings}{A vector of labelled Newick strings.}
}
\value{
A vector of labelled Newick strings corresponding to unique trees.
}
\description{
Takes as input a vector of labelled Newick strings and returns only those pertaining to unique trees.
}
\details{
Principle of free rotation.
}
\examples{

# Get information for trees with N tips:
UniqueNewickStrings(c("(A,(B,C),(D,E));", "(A,(E,D),(C,B));"))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
